refactor: move IsQrIban to an extension method to keep the Iban type API clean. Added to changelog
