+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v5.1.0
+
+- Added `iban.IsQrIban()` extension method. A Swiss or Liechtenstein QR-IBAN must have a valid QR-IID, i.e. the bank number must be within the [30000, 31999] range.
+
 ## v5.0.0
 
 ## IbanNet
